make "Update and Setup" the default action for installed packages
Also:

Repair the `dialog%` version of the GUI to disable buttons while
installation is in progress.

Repair the non-`dialog%` version of the GUI to warn a user when
they try to close a window and a package-management action is in
progress.

Update the documentation to render links as it should.